Extend modular structure with enhanced resource management, script execution, and network request tracking
- Introduced `DocumentResourceScanner` and `DocumentResources` to detect and organize scripts/stylesheets within documents.
- Enhanced `JavaScriptEngine` with lazy script execution, global context sharing, and execution of both inline and external sources.
- Updated `PageLoader` to dispatch `NetworkRequestEvent`s for finer-grained network activity monitoring, including redirects, failures, cancellations, and resource types.
- Added new resource type management (`StyleSheetResource`, `ScriptResource`) for improved clarity and modularity in loading assets.
- Expanded tests to verify lazy loading, script execution order, resource scanning, and network logging behaviors.

+++ engine-css/src/main/java/com/browicy/engine/css/CascadePriority.java
@@ -3,8 +3,7 @@ package com.browicy.engine.css;
 import com.browicy.engine.selectors.Specificity;
 import java.util.Objects;
 
-/** Vergleichsgewicht einer Deklaration innerhalb der unterstützten CSS-Kaskade. */
-public record CascadePriority(boolean inlineStyle, Specificity specificity, int sourceOrder)
+public record CascadePriority(boolean inlineStyle, Specificity specificity, long sourceOrder)
         implements Comparable<CascadePriority> {
 
     public CascadePriority {
@@ -21,6 +20,6 @@ public record CascadePriority(boolean inlineStyle, Specificity specificity, int 
         if (comparison != 0) {
             return comparison;
         }
-        return Integer.compare(sourceOrder, other.sourceOrder);
+        return Long.compare(sourceOrder, other.sourceOrder);
     }
 }
